sam/api-pattern2/storefinder-api/requirements.txt
aws-psycopg2
boto3